Refactor: navBar , Layout 추가
- 업체 측 navBar, 사용자(Extra) navBar 분리
- Layout 통일하기 위해 Layout 컴포넌트 추가
